keep shared workflow and divide readme
