+++ demo/visualization/index.html
@@ -9,6 +9,6 @@
 
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="/src/viz-browser.ts"></script>
   </body>
 </html>
