+++ demo/visualization/src/workflow.ts
@@ -1,5 +1,4 @@
-import { createWorkflow, workflowEvent } from "@llamaindex/workflow-core";
-import { withDrawing } from "@llamaindex/workflow-viz";
+import { Workflow, workflowEvent } from "@llamaindex/workflow-core";
 
 //#region define workflow events
 const startEvent = workflowEvent<string>({
@@ -25,36 +24,35 @@ const stopEvent = workflowEvent<string>({
 });
 //#endregion
 
-//#region defines workflow
-const workflow = withDrawing(createWorkflow());
+export function setupWorkflowEvents(workflow: Workflow) {
+  workflow.handle([startEvent], async (ctx) => {
+    const { sendEvent, stream } = ctx;
+    // emit 3 different events, handled separately
 
-workflow.handle([startEvent], async (ctx) => {
-  const { sendEvent, stream } = ctx;
-  // emit 3 different events, handled separately
+    sendEvent(branchAEvent.with("Branch A"));
+    sendEvent(branchBEvent.with("Branch B"));
+    sendEvent(branchCEvent.with("Branch C"));
 
-  sendEvent(branchAEvent.with("Branch A"));
-  sendEvent(branchBEvent.with("Branch B"));
-  sendEvent(branchCEvent.with("Branch C"));
+    const results = await stream.filter(branchCompleteEvent).take(3).toArray();
 
-  const results = await stream.filter(branchCompleteEvent).take(3).toArray();
+    return allCompleteEvent.with(results.map((e) => e.data).join(", "));
+  });
 
-  return allCompleteEvent.with(results.map((e) => e.data).join(", "));
-});
+  workflow.handle([branchAEvent], (_context1, branchA) => {
+    return branchCompleteEvent.with(branchA.data);
+  });
 
-workflow.handle([branchAEvent], (_context1, branchA) => {
-  return branchCompleteEvent.with(branchA.data);
-});
+  workflow.handle([branchBEvent], (_context2, branchB) => {
+    return branchCompleteEvent.with(branchB.data);
+  });
 
-workflow.handle([branchBEvent], (_context2, branchB) => {
-  return branchCompleteEvent.with(branchB.data);
-});
+  workflow.handle([branchCEvent], (_context3, branchC) => {
+    return branchCompleteEvent.with(branchC.data);
+  });
 
-workflow.handle([branchCEvent], (_context3, branchC) => {
-  return branchCompleteEvent.with(branchC.data);
-});
+  workflow.handle([allCompleteEvent], (_context4, allComplete) => {
+    return stopEvent.with(allComplete.data);
+  });
 
-workflow.handle([allCompleteEvent], (_context4, allComplete) => {
-  return stopEvent.with(allComplete.data);
-});
-
-export { workflow };
+  return workflow;
+}
